% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilly.R
\name{hillysilly}
\alias{hillysilly}
\title{Plot silly hilly}
\usage{
hillysilly(
  x = "Auckland",
  dimension = 512,
  width = 5000,
  ...,
  triangles = TRUE,
  alpha = 0.72
)
}
\arguments{
\item{x}{place names, addresses to be geocoded}

\item{dimension}{size of tiles (number of pixels each side) 1 number (will be repeated) or 2}

\item{width}{width (and height) of tiles in metres, 1 number (will be repeated) or 2}

\item{...}{ignored}

\item{triangles}{passed to \link[anglr:as.mesh3d]{anglr::as.mesh3d} (means quads are used, a literal quad primitive for every tile pixel)}

\item{alpha}{value for transparency (0,1) passed to rgl}
}
\value{
nothing, use for side-effect of a 3D plot
}
\description{
A set of input locations is derived from addresses (or placenames) in 'x'.
}
\details{
Each location is used to get elevation data in a matrix with 'dimension'
for a region around the point of 'width'. These are plotted in a purely
graphic space, with a tiling depending on how many addresses are input.
Set the dimension of 'x' to specify the tiling layout, i.e. with 'matrix(x,)'.
}
\examples{
x <- c("Hobart", "Melbourne", "Sydney", "Brisbane", "Darwin", "Perth",
"Adelaide", "Canberra")
hillysilly(matrix(x, 2L))
}
